feat(frontend): Initial framework and query functionality
This initial commit marks the beginning of our front-end development. In this commit, we've laid the foundation for our project by setting up the basic framework. Additionally, we've implemented the query functionality, allowing users to perform searches.

Please note that this is just the beginning, and more features and improvements will be added in the future.